refactor tilt series table to generic component
