+++ frontend/packages/data-portal/app/components/AccordionMetadataTable.tsx
