+++ frontend/packages/data-portal/app/components/Dataset/DatasetMetadataTable.tsx
@@ -1,8 +1,8 @@
 import { useDatasetById } from 'app/hooks/useDatasetById'
 import { i18n } from 'app/i18n'
 
-import { AccordionMetadataTable } from './AccordionMetadataTable'
-import { getTableData } from './utils'
+import { AccordionMetadataTable } from '../AccordionMetadataTable'
+import { getTableData } from '../utils'
 
 export function DatasetMetadataTable() {
   const { dataset } = useDatasetById()
